Refactor: Improve compass angle calculation and file structure
