+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/di/faithViewModelModule.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.IO
 import net.thechance.mena.faith.presentation.feature.main.MainViewModel
 import net.thechance.mena.faith.presentation.feature.qiblah.compass.CompassViewModel
 import net.thechance.mena.faith.presentation.feature.quran.bookmark.BookmarkViewModel
-import net.thechance.mena.faith.presentation.feature.quran.qiblah.calibratedevice.CalibrateDeviceViewModel
+import net.thechance.mena.faith.presentation.feature.qiblah.calibratedevice.CalibrateDeviceViewModel
 import net.thechance.mena.faith.presentation.feature.quran.search.SearchViewModel
 import net.thechance.mena.faith.presentation.feature.quran.search.args.ISearchArgs
 import net.thechance.mena.faith.presentation.feature.quran.search.args.SearchArgsImpl
